fixed bugs

+++ Main.py
@@ -41,43 +41,84 @@ music_thub = ""
 music_title=""
 yt = ""
 url = ""
+music_title_label = ""
+music_thub_label = ""
 
+#booleans
+is_search = False
+is_download = False
+exists_ok = False
 
 def toplevel_screen():      
             toplevel = customtkinter.CTkToplevel(screen)
             toplevel.title("WARNING")
-            toplevel.geometry("+%d+%d" % (screen.winfo_x()+90, screen.winfo_y()+240))
+            toplevel.geometry("+%d+%d" % (screen.winfo_x()+60, screen.winfo_y()+240))
             toplevel.geometry("250x100")
             toplevel.resizable(False, False)
+            global is_search
+            is_search = False
             label = customtkinter.CTkLabel(toplevel, text="invalid Youtube link!\nTry again with a valid link :)", font=FONT_5)
             label.pack(padx=20, pady=20)
             toplevel.grab_set()
 
+def prosess_on():      
+            proses = customtkinter.CTkToplevel(screen)
+            proses.title("WARNING")
+            proses.geometry("+%d+%d" % (screen.winfo_x()+30, screen.winfo_y()+240))
+            proses.geometry("350x100")
+            proses.resizable(False, False)
 
-def do_download(url,Download_button): 
+            if is_download == True:
+                label = customtkinter.CTkLabel(proses, text="Download is already in progress", font=FONT_5)
+                label.pack(padx=20, pady=20)
+
+            elif exists_ok == True:
+                label = customtkinter.CTkLabel(proses, text="This file already exists", font=FONT_5)
+                label.pack(padx=20, pady=20) 
+                
+            else:
+                label = customtkinter.CTkLabel(proses, text="There is an ongoing process, please wait.", font=FONT_5)
+                label.pack(padx=20, pady=20)
+            proses.grab_set()
+
+
+def do_download(url): 
     try:
-        loading_label = customtkinter.CTkLabel(master=screen,text="Please wait loading...",font=FONT_1)
-        loading_label.place(x=20,y=360)
+        global exists_ok,is_download
         path_to_download = str(os.path.join(Path.home(), 'Downloads'))
-        stream = yt.streams.filter(only_audio=True).first()
-        downloaded_file = stream.download(path_to_download)
-        base, ext = os.path.splitext(downloaded_file)
-        new_file = base + '.mp3'
-        os.rename(downloaded_file, new_file)
-        loading_label.configure(text="")
-        Download_button.configure(state="normal")
-        copleted_label = customtkinter.CTkLabel(master=screen,text=f"Completed: {path_to_download}",font=FONT_5,wraplength=300)
-        copleted_label.place(x=20,y=360)
+        path_to_check = os.path.expanduser("~/Downloads")
+        music_mp3 =f"{music_title}.mp3"
+        path_exists = os.path.exists(os.path.join(path_to_check,music_mp3))
+        if path_exists == True:
+            exists_ok = True
+            is_download = False
+            prosess_on()
+        else:
+            loading_label = customtkinter.CTkLabel(master=screen,text="Please wait, Downloading...",font=FONT_1)
+            loading_label.place(x=20,y=360)
+            stream = yt.streams.filter(only_audio=True).first()
+            downloaded_file = stream.download(path_to_download)
+            base, ext = os.path.splitext(downloaded_file)
+            new_file = base + '.mp3'
+            os.rename(downloaded_file, new_file)
+            loading_label.configure(text="")
+            is_download = False
+            copleted_label = customtkinter.CTkLabel(master=screen,text=f"Completed: {path_to_download}",font=FONT_5,wraplength=300)
+            copleted_label.place(x=20,y=360)
     except Exception as e:
         print(f"Download error: {e}") 
 
-def start_download_thread(Download_button):
-    Download_button.configure(state="disabled")
-    thread = threading.Thread(target=do_download, args=(url,))
-    thread.start()
+def start_download_thread():
+    global is_download
+    if is_download == True:
+         prosess_on()
+    else:
+        is_download = True
+        thread = threading.Thread(target=do_download, args=(url,))
+        thread.start()
 
-def about_music(Download_button):
-      global url
+def about_music():
+      global music_title_label, music_thub_label,url
       url = music_thub
       image = Image.open(requests.get(url, stream=True).raw)
       new_size = (170, 140)  
@@ -110,8 +151,6 @@ def about_music(Download_button):
       Download_button.place(x=180,y=310)
 
 def do_search():
-    searching_label = customtkinter.CTkLabel(master=screen,text="Please wait Searching...",font=FONT_1)
-    searching_label.place(x=20,y=250)
     global url
     music_link = link_entry.get()
     if music_link == "":
@@ -120,6 +159,8 @@ def do_search():
         link_entry.delete(0, END)
         toplevel_screen()
     else:
+        searching_label = customtkinter.CTkLabel(master=screen,text="Please wait, Searching...",font=FONT_1)
+        searching_label.place(x=20,y=360)
         link_entry.delete(0, END)
         global yt
         yt = YouTube(music_link)
@@ -127,11 +168,20 @@ def do_search():
         music_thub = yt.thumbnail_url
         music_title = yt.title
         searching_label.configure(text="")
+        global is_search
+        is_search = False
         about_music()
 
 def search_thread():
-    thread = threading.Thread(target=do_search)
-    thread.start()
+    global is_search, is_download
+    if is_search == True:
+        prosess_on()
+    elif is_download == True:
+        prosess_on()
+    else:
+        is_search = True
+        thread = threading.Thread(target=do_search)
+        thread.start()
 
 
 youtube_label = customtkinter.CTkLabel(master=screen,text="Youtube",font=FONT_1,text_color="green")
